+++ lib/features/kana/presentation/providers/kana_filter_provider.g.dart
@@ -1,0 +1,27 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'kana_filter_provider.dart';
+
+// **************************************************************************
+// RiverpodGenerator
+// **************************************************************************
+
+String _$kanaCategoryFilterHash() =>
+    r'e57d1a0a7a6832be7130be89c941e917bd699d10';
+
+/// See also [KanaCategoryFilter].
+@ProviderFor(KanaCategoryFilter)
+final kanaCategoryFilterProvider =
+    AutoDisposeNotifierProvider<KanaCategoryFilter, KanaCategory>.internal(
+      KanaCategoryFilter.new,
+      name: r'kanaCategoryFilterProvider',
+      debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
+          ? null
+          : _$kanaCategoryFilterHash,
+      dependencies: null,
+      allTransitiveDependencies: null,
+    );
+
+typedef _$KanaCategoryFilter = AutoDisposeNotifier<KanaCategory>;
+// ignore_for_file: type=lint
+// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member, deprecated_member_use_from_same_package